Add checking webOSRelaunch and remove starting the animation at the first launch
:Release Notes:
Add checking webOSRelaunch and remove starting the animation at the first launch

:Detailed Notes:
If the event webOSRelaunch is received,
the app should be show icons with animation.
Because the app is running in background,
it is launched with the webOSRelaunch event.

At the first launch time, the animation is not started.
After getting webOSRelaunch event, the animation is shown.

:Testing Performed:
See the CCC JIRA ticket.

:QA Notes:
NA

:Issues Addressed:
[PLAT-86297] CCC: Introduce Home/Volume UI for webOS Auto reference UX
[PLAT-85708] Home app - CCC into the OSE and webOS Auto builds
[PLAT-76652] [App] Launcher - requirements for OSE 2.0
[PLAT-86421] Not able to relaunch "com.webos.app.home" with keepAlive launch param

Change-Id: Id71ff718fffe5b8a94d8fe2f43e9d123a55bed82